TCS-UI developer 2020
In this assignment, the task was that given a std ip, calculate the minimum value of sum of each elements of  | A[i]-B[i] |
1st line contains the no. of test cases, then comes the size of the array, and then came the arrays in each of the following lines

The main logic behind this program is that the sum will be minimum, when each of its components is minimum, which is possible only when we are substracting greatest no. of one array with greatest no. of another array